+++ models/tt_dit/experimental/tests/conftest.py
@@ -1,0 +1,12 @@
+# SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
+
+# SPDX-License-Identifier: Apache-2.0
+
+import os
+
+import torch
+from loguru import logger
+
+num_torch_threads = max(1, os.cpu_count())
+logger.info(f"Setting torch num_threads to {num_torch_threads}")
+torch.set_num_threads(num_torch_threads)
